Groundstation/Plotting/QTgraph/pyproject.toml
[tool.poetry]
name = "gs-monitor"
version = "0.1.0"
description = ""
authors = ["Fynn <f.gewiese@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
opengl-registry = "^0.3.1"
serial = "^0.0.97"
pyqtgraph = "^0.13.7"
pyqt5 = "=5.15.11"   
pyqt5-qt5 = "=5.15.2"
pyopengl = "^3.1.7"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pyqt5 = {version = "^5.15.11", optional = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"